refactored get by category method

+++ src/main/java/dev/yerokha/cookscorner/entity/Category.java
@@ -20,7 +20,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private Long categoryId;
+    private Byte categoryId;
 
     @Column(name = "category_name")
     private String categoryName;
